Add small/large seed size and ability to preserve seed data after tests

+++ src/DotnetTests/Fixtures/FilledApplicationDbContextFixture.cs
@@ -11,6 +11,10 @@ public class FilledApplicationDbContextFixture : IAsyncLifetime
 
     private readonly TestSeeder _testSeeder;
 
+    private readonly bool _preserveSeededData;
+
+    private readonly bool _envIsDev;
+
     public FilledApplicationDbContextFixture()
     {
         var envFilePath = Directory.GetCurrentDirectory() + "/../../../.env";
@@ -21,6 +25,13 @@ public class FilledApplicationDbContextFixture : IAsyncLifetime
                 string? line;
                 while ((line = reader.ReadLine()) != null)
                 {
+                    if (
+                        line.Count() == 0
+                        || line.TrimStart().FirstOrDefault() == '#'
+                    )
+                    {
+                        continue;
+                    }
                     int i = line.IndexOf("=");
                     if (i == -1)
                     {
@@ -35,28 +46,43 @@ public class FilledApplicationDbContextFixture : IAsyncLifetime
                 }
             }
         }
+        _preserveSeededData =
+            Environment.GetEnvironmentVariable("PRESERVE_SEEDED_DATA")
+            == "true";
+        _envIsDev = Environment.GetEnvironmentVariable("ENV") == "dev";
+        string connectionString = _envIsDev
+            ? "LOCAL_DB_CONNECTION_STRING"
+            : "TEST_DB_CONNECTION_STRING";
         var options = new DbContextOptionsBuilder<ApplicationDbContext>()
-            .UseNpgsql(
-                Environment.GetEnvironmentVariable("TEST_DB_CONNECTION_STRING")
-            )
+            .UseNpgsql(Environment.GetEnvironmentVariable(connectionString))
             //.EnableSensitiveDataLogging()
             //.LogTo(Console.WriteLine)
             .Options;
 
         ApplicationDbContext c = new ApplicationDbContext(options);
-        c.Database.Migrate();
+        if (!c.Database.CanConnect())
+        {
+            c.Database.Migrate();
+        }
         context = c;
         _testSeeder = new TestSeeder(context, UserStoreTests.GetUserStore());
     }
 
     public async Task InitializeAsync()
     {
-        await _testSeeder.Seed(TestSeeder.Small);
+        string seedSize = Environment.GetEnvironmentVariable("LARGE_SEED_SIZE")
+            is null
+            ? TestSeeder.Small
+            : TestSeeder.Large;
+        await _testSeeder.Seed(seedSize);
     }
 
     public async Task DisposeAsync()
     {
-        context.Database.EnsureDeleted();
+        if (!_preserveSeededData)
+        {
+            context.Database.EnsureDeleted();
+        }
         await context.DisposeAsync();
     }
 }
